+++ planck-cljs/test/planck/requiring_resolve_ns.cljs
@@ -1,0 +1,3 @@
+(ns planck.requiring-resolve-ns)
+
+(def a 3)