+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 ### Added
-- Support setting user agent in `planck.http` ([#839](https://github.com/mfikes/planck/issues/839))
+- Support setting user agent in `planck.http` ([#838](https://github.com/mfikes/planck/issues/838))
 
 ### Fixed
 - Make s/explain provide names of core fns ([#832](https://github.com/mfikes/planck/issues/832))